+++ npc/character/tags/__init__.py
@@ -8,3 +8,4 @@ from .unknown_tag import UnknownTag
 from .description_tag import DescriptionTag
 from .group_tag import GroupTag
 from .tag_container import TagContainer
+from .sub_tag import SubTag